chore!: Refactor Teamcity integration
